requirements.txt
Flask==2.2.2
Flask-SQLAlchemy==2.5.1
psycopg2-binary==2.9.6
python-dotenv==1.0.0
Werkzeug==2.2.3
Flask-Cors==3.0.10
PyJWT==2.8.0
